requirements/base.txt
Django==1.7.2
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
psycopg2==2.7.5
Pillow==2.6.1
django-multiupload==0.3
django-endless-pagination==2.0
python-dateutil==2.4.0
django-messages==0.5.1
django-templated-email==0.4.9
celery==3.1.17
sorl-thumbnail==12.2
pytz==2014.10
